requirements.txt
pyusb>=1.2.1
pyserial>=3.5
lxml>=4.9.0
requests>=2.28.0
